Refactor to fix SKIP_TABLES=true compilation
