+++ osquery/sql/hash.h
